rename package aidapta to visarchpy
