+++ src/visarchpy/edge_detection/detection.py
