+++ src/visarchpy/pipelines/dinov2.py
